Añadimos seguridad al proceso de borrado

+++ delete-templates.php
@@ -54,6 +54,11 @@ function delete_themes_options_page_html()
     $themes = delete_themes_get_list();
 
     if (isset($_GET[DELETE_THEMES_PARAM])) {
+
+        if (!wp_verify_nonce($_REQUEST['nonce'], $_GET[DELETE_THEMES_PARAM])) {
+            wp_die("Operación no permitida");
+        }
+
         if (delete_themes_execute($_GET[DELETE_THEMES_PARAM], $themes)) {
             $url = DELETE_THEMES_URL;
             echo "<script>location.href = '$url';</script>";
